Add support for convert/thumb options job format
Dragonfly supports specifying the format and frame as a hash. This change
adds support for existing urls using this syntax for both convert and
thumb.

Example job created by dragonfly 1.0.7:
```ruby
[["f", "7a7/9fb/7a79fb0c87f4db253eb46aac26553351.jpg"],
 ["p", :thumb, "100x100", {:format=>"png", :frame=>1}]]
```

This also adds a stringifier, which handles the creation of the unique
steps string that is used for signing the url. The previous
implementation could not handle structs and the required
canonicalization. This code mirrors the dragonfly gem's
to_unique_dragonfly_s core class extensions.